requirements.txt
absl-py==2.3.1
ansitable==0.11.5
arm_pytorch_utilities==0.4.3
asttokens==3.0.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.5.0
certifi==2025.11.12
cfgv==3.4.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
colored==1.4.4
comm==0.2.3
contourpy==1.1.1
cv-bridge==1.13.0.post0  # 若在纯非ROS环境请删去；已在上面ROS清单
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.16
decorator==5.2.1
distlib==0.4.0
distro==1.7.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.20.1
empy==4.2
etils==1.3.0
eval_type_backport==0.2.2
evdev==1.9.2
exceptiongroup==1.3.1
executing==2.2.1
Farama-Notifications==0.0.4
fast_kinematics==0.2.2
feetech-servo-sdk==1.0.0
filelock==3.16.1
flatbuffers>=24.3.0
fonttools==4.57.0
fsspec==2025.3.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
glfw==2.9.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
gymnasium==0.29.1
h5py==3.11.0
hf-xet==1.2.0
hid==1.0.4
hidapi==0.14.0.post4
huggingface-hub==0.36.0
identify==2.6.1
idna==3.3
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==4.6.4
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.13.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
mani_skill==3.0.0.dev13
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.3.2
mplib==0.1.1
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
num2words==0.5.14
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
parso==0.8.5
pexpect==4.8.0
pgraph-python==0.6.3
pickleshare==0.7.5
pillow==9.0.1
platformdirs==4.3.6
pre-commit==3.5.0
progress==1.6.1
prompt_toolkit==3.0.52
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pyglm==2.8.2
Pygments==2.19.2
pynput==1.7.7
pynvml==12.0.0
PyOpenGL==3.1.9
pyparsing==3.3.2
pyperclip==1.11.0
pyrealsense2==2.55.1.6486
pyserial==3.5
python-dateutil==2.9.0.post0
python-xlib==0.33
pytorch-kinematics==0.7.6
pytorch-seed==0.2.0
pyttsx3==2.99
PyYAML==5.4.1
pyzmq==27.0.1
requests==2.25.1
requests-oauthlib==2.0.0
rich==14.1.0
rlds==0.1.8
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.8
sapien==3.0.2
scipy==1.15.3
shtab==1.7.2
six==1.16.0
smmap==5.0.2
spatialgeometry==1.1.0
spatialmath-python==1.1.14
stack-data==0.6.3
swift-sim==1.1.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.16.1
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
toppra==0.6.3
torch==2.10.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.11.1
triton==3.4.0
typeguard==4.4.4
typing_extensions==4.15.0
tyro==1.0.5
urllib3==1.26.5
virtualenv==20.33.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.0.6
wrapt==1.17.2
xarm-python-sdk==1.15.3
zipp==1.0.0
